Remove legacy piece generation paths
