+++ packages/llm/src/index.ts
@@ -7,8 +7,5 @@ export * from "./prompts/json-import.ts";
 export * from "./prompts/prompting.ts";
 export * from "./prompts/pattern-fix.ts";
 export * from "./prompts/pattern-guide.ts";
-export * from "./prompts/spec-and-schema-gen.ts";
-export * from "./prompts/code-and-schema-gen.ts";
-export * from "./prompts/workflow-classification.ts";
 export * from "./prompts/prompting.ts";
 export * from "./options.ts";
