--- packages/piece/src/imagine.ts
@@ -1,24 +1,0 @@
-import { Cell } from "@commonfabric/runner";
-
-// Re-export workflow types and functions from workflow module
-export type { WorkflowConfig, WorkflowType } from "./workflow.ts";
-
-/**
- * Structure representing a successfully parsed mention in a prompt
- */
-export interface ParsedMention {
-  id: string;
-  name: string;
-  originalText: string;
-  startIndex: number;
-  endIndex: number;
-  piece: Cell<unknown>;
-}
-
-/**
- * Result of processing a prompt with mentions
- */
-export interface ProcessedPrompt {
-  text: string; // Processed text with mentions replaced by readable names
-  mentions: Record<string, Cell<unknown>>; // Map of mention IDs to piece cells
-}
