+++ packages/piece/src/index.ts
@@ -1,14 +1,1 @@
 export { getPatternIdFromPiece, pieceId, PieceManager } from "./manager.ts";
-export { extractUserCode } from "./iframe/static.ts";
-export { createDataPiece } from "./data-piece.ts";
-export { type ParsedMention, type ProcessedPrompt } from "./imagine.ts";
-
-// Export workflow module
-export {
-  type IntentClassificationResult,
-  processWorkflow,
-  type WorkflowConfig,
-  type WorkflowForm,
-  WORKFLOWS,
-  type WorkflowType,
-} from "./workflow.ts";
